+++ packages/http.go
@@ -15,11 +15,34 @@ import (
 	"github.com/elastic/package-registry/util"
 )
 
+// ServePackage is used by artifactsHandler to serve packages and signatures.
 func ServePackage(w http.ResponseWriter, r *http.Request, p *Package) {
 	span, _ := apm.StartSpan(r.Context(), "ServePackage", "app")
 	defer span.End()
 
-	packagePath := p.BasePath
+	if p.RemoteResolver() != nil {
+		p.RemoteResolver().RedirectArtifactsHandler(w, r, p)
+		return
+	}
+	serveLocalPackage(w, r, p, p.BasePath)
+}
+
+// ServePackageSignature is used by signaturesHandler to serve signatures.
+func ServePackageSignature(w http.ResponseWriter, r *http.Request, p *Package) {
+	span, _ := apm.StartSpan(r.Context(), "ServePackageSignature", "app")
+	defer span.End()
+
+	if p.RemoteResolver() != nil {
+		p.RemoteResolver().RedirectSignaturesHandler(w, r, p)
+		return
+	}
+	serveLocalPackage(w, r, p, p.BasePath+".sig")
+}
+
+func serveLocalPackage(w http.ResponseWriter, r *http.Request, p *Package, packagePath string) {
+	span, _ := apm.StartSpan(r.Context(), "ServePackage", "app")
+	defer span.End()
+
 	logger := util.Logger().With(zap.String("file.name", packagePath))
 
 	f, err := os.Stat(packagePath)
@@ -29,9 +52,8 @@ func ServePackage(w http.ResponseWriter, r *http.Request, p *Package) {
 		return
 	}
 
-	w.Header().Set("Content-Type", "application/gzip")
-
 	if f.IsDir() {
+		w.Header().Set("Content-Type", "application/zip")
 		err = archiver.ArchivePackage(w, archiver.PackageProperties{
 			Name:    p.Name,
 			Version: p.Version,
@@ -39,18 +61,24 @@ func ServePackage(w http.ResponseWriter, r *http.Request, p *Package) {
 		})
 		if err != nil {
 			logger.Error("archiving package path failed", zap.Error(err))
-			return
 		}
-	} else {
-		http.ServeFile(w, r, packagePath)
+		return
 	}
+
+	http.ServeFile(w, r, packagePath)
 }
 
-func ServeFile(w http.ResponseWriter, r *http.Request, p *Package, name string) {
+// ServePackageResource is used by staticHandler.
+func ServePackageResource(w http.ResponseWriter, r *http.Request, p *Package, packageFilePath string) {
 	span, _ := apm.StartSpan(r.Context(), "ServePackage", "app")
 	defer span.End()
 
-	logger := util.Logger().With(zap.String("file.name", name))
+	if p.RemoteResolver() != nil {
+		p.RemoteResolver().RedirectStaticHandler(w, r, p, packageFilePath)
+		return
+	}
+
+	logger := util.Logger().With(zap.String("file.name", packageFilePath))
 
 	fs, err := p.fs()
 	if os.IsNotExist(err) {
@@ -63,7 +91,7 @@ func ServeFile(w http.ResponseWriter, r *http.Request, p *Package, name string) 
 		return
 	}
 
-	stat, err := fs.Stat(name)
+	stat, err := fs.Stat(packageFilePath)
 	if os.IsNotExist(err) {
 		http.Error(w, "resource not found", http.StatusNotFound)
 		return
@@ -74,7 +102,7 @@ func ServeFile(w http.ResponseWriter, r *http.Request, p *Package, name string) 
 		return
 	}
 
-	f, err := fs.Open(name)
+	f, err := fs.Open(packageFilePath)
 	if err != nil {
 		logger.Error("failed to open file", zap.Error(err))
 		http.Error(w, "internal server error", http.StatusInternalServerError)
@@ -82,9 +110,5 @@ func ServeFile(w http.ResponseWriter, r *http.Request, p *Package, name string) 
 	}
 	defer f.Close()
 
-	http.ServeContent(w, r, name, stat.ModTime(), f)
-}
-
-func ServeSignature(w http.ResponseWriter, r *http.Request, p *Package) {
-	http.ServeFile(w, r, p.BasePath+".sig")
+	http.ServeContent(w, r, packageFilePath, stat.ModTime(), f)
 }