+++ static.go
@@ -50,8 +50,7 @@ func staticHandler(indexer Indexer, cacheTime time.Duration) http.HandlerFunc {
 		}
 
 		cacheHeaders(w, cacheTime)
-
-		packages.ServeFile(w, r, packageList[0], params.fileName)
+		packages.ServePackageResource(w, r, packageList[0], params.fileName)
 	}
 }